Move the `FeeHistoryResult` type to the eth-rpc
