+++ substrate/frame/revive/rpc/src/fee_history_provider.rs
@@ -14,8 +14,8 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-use crate::{ClientError, client::SubstrateBlockNumber};
-use pallet_revive::evm::{Block, FeeHistoryResult, ReceiptInfo};
+use crate::{ClientError, FeeHistoryResult, client::SubstrateBlockNumber};
+use pallet_revive::evm::{Block, ReceiptInfo};
 use sp_core::U256;
 use std::{collections::BTreeMap, sync::Arc};
 use tokio::sync::RwLock;
