+++ substrate/frame/revive/rpc/src/types/fee_history.rs
@@ -1,0 +1,45 @@
+// This file is part of Substrate.
+
+// Copyright (C) Parity Technologies (UK) Ltd.
+// SPDX-License-Identifier: Apache-2.0
+
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//  http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use crate::*;
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
+#[serde(rename_all = "camelCase")]
+pub struct FeeHistoryResult {
+	/// Lowest number block of the returned range.
+	pub oldest_block: U256,
+
+	/// An array of block base fees per gas.
+	///
+	/// This includes the next block after the newest of the returned range, because this value can
+	/// be derived from the newest block. Zeroes are returned for pre-EIP-1559 blocks.
+	#[serde(default, skip_serializing_if = "Vec::is_empty")]
+	pub base_fee_per_gas: Vec<U256>,
+
+	/// An array of block gas used ratios.
+	/// These are calculated as the ratio of `gasUsed` and `gasLimit`.
+	pub gas_used_ratio: Vec<f64>,
+
+	/// A two-dimensional array of effective priority fees per gas at the requested block
+	/// percentiles.
+	///
+	/// A given percentile sample of effective priority fees per gas from a single block in
+	/// ascending order, weighted by gas used. Zeroes are returned if the block is empty.
+	#[serde(default, skip_serializing_if = "Vec::is_empty")]
+	pub reward: Vec<Vec<U256>>,
+}
